backend/requirements.txt
flask
python-dotenv
flask_cors
openai
psycopg2-binary